fps values are now floats instead of ints

+++ facefusion/globals.py
@@ -44,7 +44,7 @@ output_image_quality : Optional[int] = None
 output_video_encoder : Optional[OutputVideoEncoder] = None
 output_video_preset : Optional[OutputVideoPreset] = None
 output_video_quality : Optional[int] = None
-output_video_fps : Optional[int] = None
+output_video_fps : Optional[float] = None
 skip_audio : Optional[bool] = None
 # frame processors
 frame_processors : List[str] = []
